backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.3.1
Pillow==10.4.0
